+++ lib/presentation/widgets/dialogs.dart
@@ -21,12 +21,12 @@ quizCompletionDialog(BuildContext context) {
       return BlocBuilder<AppThemeCubit, AppThemeState>(
         builder: (context, themeState) {
           return Dialog(
-              backgroundColor: (themeState as AppThemeSet).themeClass.buttonBackgroundColor2,
+              backgroundColor: (themeState as AppThemeSet).themeClass.quizBackgroundColor,
             shape: RoundedRectangleBorder(
                 borderRadius: BorderRadius.circular(10.sp)),
             child: SizedBox(
               width: 90.w,
-              height: 30.h,
+              height: 29.h,
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.start,
                 children: [
@@ -36,10 +36,11 @@ quizCompletionDialog(BuildContext context) {
                       Flexible(
                         child: Padding(
                           padding: EdgeInsets.all(5.w),
-                          child: const CustomText(
+                          child: CustomText(
                             textString: 'Do you want to submit your answers?',
                             fontWeight: FontWeight.bold,
                             textAlign: TextAlign.center,
+                            textColor: themeState.themeClass.white,
                           ),
                         ),
                       ),
@@ -51,6 +52,7 @@ quizCompletionDialog(BuildContext context) {
                     child: CustomTextField(
                       controller: ctrl,
                       hintText: 'Enter your name...',
+                      fillColor: themeState.themeClass.formFieldBackgroundColor,
                     ),
                   ),
 
@@ -59,7 +61,7 @@ quizCompletionDialog(BuildContext context) {
                     child: CustomButton(
                       buttonColor: themeState.themeClass.quizButtonBackgroundColor,
                       buttonText: 'submit'.toUpperCase(),
-                      buttonTextColor: themeState.themeClass.textColor_2,
+                      buttonTextColor: themeState.themeClass.white,
                       buttonSize: 90.w,
                       borderRadius: 15,
                       fontWeight: FontWeight.bold,